Refactored url shortener to support updated git.io api
